chore: rename TextInput and SelectInput to index.tsx
