+++ src/components/MultiStepForm/Steps/DiversityInformation.tsx
@@ -2,7 +2,7 @@ import * as Yup from "yup";
 
 import Step from "../Step";
 import { sleep } from "../../../utils";
-import SelectInput from "../../SelectInput/SelectInput";
+import SelectInput from "../../SelectInput";
 import { colorOptions, disabilityOptions } from "../../../lib/constants";
 
 const diversityInformationSchema = Yup.object({
