+++ src/components/index.ts
@@ -3,7 +3,7 @@ export { default as MultiStepForm } from "./MultiStepForm";
 export { default as DesktopIllustration } from "./DesktopIllustration";
 export { default as StepsController } from "./MultiStepForm/StepsController";
 export { default as ErrorMessage } from "./ErrorMessage";
-export { default as TextInput } from "./TextInput/TextInput";
+export { default as TextInput } from "./TextInput";
 export { default as RadioInput } from "./RadioInput";
 export { default as CheckboxInput } from "./CheckboxInput";
-export { default as SelectInput } from "./SelectInput/SelectInput";
+export { default as SelectInput } from "./SelectInput";
